+++ package.json
@@ -13,8 +13,7 @@
     "check-types": "turbo check-types",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "ui:add": "node scripts/ui-add.js",
-    "secrets": "node scripts/setup-secrets.js",
-    "pull-secrets": "node scripts/pull-secrets.js"
+    "pull-secrets": "node scripts/setup-secrets.js"
   },
   "devDependencies": {
     "@uwdsc/eslint-config": "workspace:*",
